added packages and create test class
