+++ src/TestEuro.java
@@ -1,0 +1,33 @@
+import org.junit.Test;
+
+public class TestEuro {
+    @Test
+    public void testGetValore() {
+
+    }
+
+    @Test
+    public void testMinoreDi() {
+
+    }
+
+    @Test
+    public void testSomma() {
+
+    }
+
+    @Test
+    public void testSottrai() {
+
+    }
+
+    @Test
+    public void testStampa() {
+
+    }
+
+    @Test
+    public void testUgualeA() {
+
+    }
+}
