+++ src/code/Business_logic/Account.java
@@ -1,4 +1,4 @@
-package code;
+package code.Business_logic;
 // Account.java
 // Represents a bank account
