+++ src/code/Business_logic/CashDispenser.java
@@ -1,4 +1,4 @@
-package code;
+package code.Business_logic;
 // CashDispenser.java
 // Represents the cash dispenser of the ATM
