+++ src/code/Business_logic/Euro.java
@@ -1,4 +1,4 @@
-package code;
+package code.Business_logic;
 public class Euro {
