+++ src/code/Database/BankDatabase.java
@@ -1,6 +1,8 @@
-package code;
+package code.Database;
 // BankDatabase.java
 // Represents the bank account information database 
+
+import code.Business_logic.Account;
 
 public class BankDatabase
 {
