+++ src/code/GUI/ATMCaseStudy.java
@@ -1,24 +1,16 @@
-package code;
-// Keypad.java
-// Represents the keypad of the ATM
-import java.util.Scanner; // program uses Scanner to obtain user input
+package code.GUI;
+// ATMCaseStudy.java
+// Driver program for the ATM case study
 
-public class Keypad
+public class ATMCaseStudy
 {
-   private Scanner input; // reads data from the command line
-                         
-   // no-argument constructor initializes the Scanner
-   public Keypad()
+   // main method creates and runs the ATM
+   public static void main( String[] args )
    {
-      input = new Scanner( System.in );    
-   } // end no-argument Keypad constructor
-
-   // return an integer value entered by user 
-   public int getInput()
-   {
-      return input.nextInt(); // we assume that user enters an integer  
-   } // end method getInput
-} // end class Keypad  
+      ATM theATM = new ATM();    
+      theATM.run();
+   } // end main
+} // end class ATMCaseStudy
